+++ src/api/mod.rs
@@ -5,3 +5,4 @@ pub mod io;
 pub mod net;
 pub mod os;
 pub mod time;
+pub mod lua_logger;